[ENH] same enhancement for cnn

+++ aeon/classification/deep_learning/cnn.py
@@ -66,7 +66,24 @@ class CNNClassifier(BaseDeepClassifier):
     callbacks       : keras.callbacks, default=model_checkpoint to save best
                       model on training loss
     file_path       : file_path for the best model (if checkpoint is used as callback)
-
+    save_best_model     : bool, default = False
+        Whether or not to save the best model, if the
+        modelcheckpoint callback is used by default,
+        this condition, if True, will prevent the
+        automatic deletion of the best saved model from
+        file and the user can choose the file name
+    save_last_model     : bool, default = False
+        Whether or not to save the last model, last
+        epoch trained, using the base class method
+        save_last_model_to_file
+    best_file_name      : str, default = "best_model"
+        The name of the file of the best model, if
+        save_best_model is set to False, this parameter
+        is discarded
+    last_file_name      : str, default = "last_model"
+        The name of the file of the last model, if
+        save_last_model is set to False, this parameter
+        is discarded
 
     Notes
     -----
@@ -105,6 +122,10 @@ class CNNClassifier(BaseDeepClassifier):
         batch_size=16,
         callbacks=None,
         file_path="./",
+        save_best_model=False,
+        save_last_model=False,
+        best_file_name="best_model",
+        last_file_name="last_model",
         verbose=False,
         loss="mean_squared_error",
         metrics=None,
@@ -113,7 +134,7 @@ class CNNClassifier(BaseDeepClassifier):
         optimizer=None,
     ):
         _check_dl_dependencies(severity="error")
-        super(CNNClassifier, self).__init__()
+        super(CNNClassifier, self).__init__(last_file_name=last_file_name)
 
         self.n_layers = n_layers
         self.kernel_size = kernel_size
@@ -130,6 +151,10 @@ class CNNClassifier(BaseDeepClassifier):
         self.batch_size = batch_size
         self.callbacks = callbacks
         self.file_path = file_path
+        self.save_best_model = save_best_model
+        self.save_last_model = save_last_model
+        self.best_file_name = best_file_name
+        self.last_file_name = last_file_name
         self.verbose = verbose
         self.loss = loss
         self.metrics = metrics
@@ -222,7 +247,9 @@ class CNNClassifier(BaseDeepClassifier):
         if self.verbose:
             self.training_model_.summary()
 
-        self.file_name_ = str(time.time_ns())
+        self.file_name_ = (
+            self.best_file_name if self.save_best_model else str(time.time_ns())
+        )
 
         self.callbacks_ = (
             [
@@ -249,9 +276,13 @@ class CNNClassifier(BaseDeepClassifier):
             self.model_ = tf.keras.models.load_model(
                 self.file_path + self.file_name_ + ".hdf5", compile=False
             )
-            os.remove(self.file_path + self.file_name_ + ".hdf5")
+            if not self.save_best_model:
+                os.remove(self.file_path + self.file_name_ + ".hdf5")
         except FileNotFoundError:
             self.model_ = deepcopy(self.training_model_)
+
+        if self.save_last_model:
+            self.save_last_model_to_file(file_path=self.file_path)
 
         return self
